Move user namespace/remapped root to experimental build
This moves the UI and base configuration settings to only be accessible
from the experimental build. The uid and gid maps will be in the daemon
configuration, but will be nil, effectively turning off any support for
user namespaces in the daemon and containers run on the daemon.  The
`--root` setting is now available in the experimental build tag with the
same features as before.

Docker-DCO-1.1-Signed-off-by: Phil Estes <estesp@linux.vnet.ibm.com> (github: estesp)